fix(image): add a last layer with a poweroff state
After build the image last layer can not be in save state, because if the
image is exported the data in the save state is discarded without writing
it to the disk.
The image last layer is always POWEROFF, this guarantees that the last
layer has all data saved to the disk.